Extract common functionality from Features
