+++ common/src/main/java/com/wynntils/features/DialogueOptionOverrideFeature.java
@@ -25,8 +25,9 @@ public class DialogueOptionOverrideFeature extends FeatureBase {
     @SubscribeEvent
     public void onDialogueKeyPress(KeyInputEvent e) {
         if (!WynnUtils.onWorld() || e.getAction() != 1) return; // Only send packet on presses, not releases
+
         if (e.getKey() - 49 == McUtils.player().getInventory().selected) { // keys 1-9 are +49 offset from hotbar
-            McUtils.mc().getConnection().send(new ServerboundSetCarriedItemPacket(e.getKey() - 49));
+            McUtils.sendPacket(new ServerboundSetCarriedItemPacket(e.getKey() - 49));
         }
     }
 }
