+++ common/src/main/java/com/wynntils/features/MythicBlockerFeature.java
@@ -10,10 +10,12 @@ import com.wynntils.core.features.properties.GameplayImpact;
 import com.wynntils.core.features.properties.PerformanceImpact;
 import com.wynntils.core.features.properties.Stability;
 import com.wynntils.mc.event.InventoryKeyPressEvent;
+import com.wynntils.mc.utils.ItemUtils;
 import com.wynntils.mc.utils.McUtils;
+import com.wynntils.wc.utils.ContainerUtils;
 import com.wynntils.wc.utils.WynnUtils;
 import net.minecraft.ChatFormatting;
-import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
+import net.minecraft.core.NonNullList;
 import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.world.item.ItemStack;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -27,17 +29,14 @@ public class MythicBlockerFeature extends FeatureBase {
 
     @SubscribeEvent
     public void onChestCloseAttempt(InventoryKeyPressEvent e) {
-        if (!WynnUtils.onWorld() || !McUtils.mc().options.keyInventory.matches(e.getKeyCode(), e.getScanCode())) return;
-        if (!(McUtils.mc().screen instanceof AbstractContainerScreen<?>)) return;
-        String title = McUtils.mc().screen.getTitle().getString();
-        if (!title.startsWith("Loot Chest")
-                && !title.startsWith("Daily Rewards")
-                && !title.contains("Objective Rewards")) return;
+        if (!WynnUtils.onWorld()) return;
+        if (!McUtils.mc().options.keyInventory.matches(e.getKeyCode(), e.getScanCode())) return;
+        if (!ContainerUtils.isLootOrRewardChest(McUtils.mc().screen)) return;
 
-        AbstractContainerScreen<?> screen = (AbstractContainerScreen<?>) McUtils.mc().screen;
+        NonNullList<ItemStack> items = ContainerUtils.getItems(McUtils.mc().screen);
         for (int i = 0; i < 27; i++) {
-            ItemStack stack = screen.getMenu().getItems().get(i);
-            if (stack.getDisplayName().getString().contains(ChatFormatting.DARK_PURPLE.toString())) {
+            ItemStack stack = items.get(i);
+            if (ItemUtils.isMythic(stack)) {
                 McUtils.sendMessageToClient(new TranslatableComponent("feature.wynntils.mythicBlocker.closingBlocked")
                         .withStyle(ChatFormatting.RED));
                 e.setCanceled(true);
