+++ common/src/main/java/com/wynntils/mc/utils/RenderUtils.java
@@ -4,6 +4,8 @@
  */
 package com.wynntils.mc.utils;
 
+import com.mojang.blaze3d.pipeline.RenderTarget;
+import com.mojang.blaze3d.platform.NativeImage;
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.BufferBuilder;
 import com.mojang.blaze3d.vertex.BufferUploader;
@@ -12,7 +14,17 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import com.mojang.blaze3d.vertex.Tesselator;
 import com.mojang.blaze3d.vertex.VertexFormat;
 import com.mojang.math.Matrix4f;
+import java.awt.Image;
+import java.awt.Toolkit;
+import java.awt.datatransfer.DataFlavor;
+import java.awt.datatransfer.Transferable;
+import java.awt.datatransfer.UnsupportedFlavorException;
+import java.awt.image.BufferedImage;
+import java.util.List;
+import net.minecraft.client.gui.Font;
+import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
 import net.minecraft.client.renderer.GameRenderer;
+import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.resources.ResourceLocation;
 
 public class RenderUtils {
@@ -163,5 +175,211 @@ public class RenderUtils {
         bufferBuilder.end();
         BufferUploader.end(bufferBuilder);
         RenderSystem.disableBlend();
+    }
+
+    public static void fillGradient(
+            Matrix4f matrix,
+            BufferBuilder builder,
+            int x1,
+            int y1,
+            int x2,
+            int y2,
+            int blitOffset,
+            int colorA,
+            int colorB) {
+        float f = (float) (colorA >> 24 & 0xFF) / 255.0f;
+        float g = (float) (colorA >> 16 & 0xFF) / 255.0f;
+        float h = (float) (colorA >> 8 & 0xFF) / 255.0f;
+        float i = (float) (colorA & 0xFF) / 255.0f;
+        float j = (float) (colorB >> 24 & 0xFF) / 255.0f;
+        float k = (float) (colorB >> 16 & 0xFF) / 255.0f;
+        float l = (float) (colorB >> 8 & 0xFF) / 255.0f;
+        float m = (float) (colorB & 0xFF) / 255.0f;
+        builder.vertex(matrix, x2, y1, blitOffset).color(g, h, i, f).endVertex();
+        builder.vertex(matrix, x1, y1, blitOffset).color(g, h, i, f).endVertex();
+        builder.vertex(matrix, x1, y2, blitOffset).color(k, l, m, j).endVertex();
+        builder.vertex(matrix, x2, y2, blitOffset).color(k, l, m, j).endVertex();
+    }
+
+    public static void drawTooltip(List<ClientTooltipComponent> lines, PoseStack poseStack, Font font) {
+        int tooltipWidth = 0;
+        int tooltipHeight = lines.size() == 1 ? -2 : 0;
+
+        for (ClientTooltipComponent clientTooltipComponent : lines) {
+            int lineWidth = clientTooltipComponent.getWidth(font);
+            if (lineWidth > tooltipWidth) {
+                tooltipWidth = lineWidth;
+            }
+            tooltipHeight += clientTooltipComponent.getHeight();
+        }
+
+        // background box
+        poseStack.pushPose();
+        int tooltipX = 4;
+        int tooltipY = 4;
+        // somewhat hacky solution to get around transparency issues - these colors were chosen to best match
+        // how tooltips are displayed in-game
+        int backgroundColor = 0xFF100010;
+        int borderColorStart = 0xFF25005B;
+        int borderColorEnd = 0xFF180033;
+        int zLevel = 400;
+        Tesselator tesselator = Tesselator.getInstance();
+        BufferBuilder bufferBuilder = tesselator.getBuilder();
+        RenderSystem.setShader(GameRenderer::getPositionColorShader);
+        bufferBuilder.begin(VertexFormat.Mode.QUADS, DefaultVertexFormat.POSITION_COLOR);
+        Matrix4f matrix4f = poseStack.last().pose();
+        fillGradient(
+                matrix4f,
+                bufferBuilder,
+                tooltipX - 3,
+                tooltipY - 4,
+                tooltipX + tooltipWidth + 3,
+                tooltipY - 3,
+                zLevel,
+                backgroundColor,
+                backgroundColor);
+        fillGradient(
+                matrix4f,
+                bufferBuilder,
+                tooltipX - 3,
+                tooltipY + tooltipHeight + 3,
+                tooltipX + tooltipWidth + 3,
+                tooltipY + tooltipHeight + 4,
+                zLevel,
+                backgroundColor,
+                backgroundColor);
+        fillGradient(
+                matrix4f,
+                bufferBuilder,
+                tooltipX - 3,
+                tooltipY - 3,
+                tooltipX + tooltipWidth + 3,
+                tooltipY + tooltipHeight + 3,
+                zLevel,
+                backgroundColor,
+                backgroundColor);
+        fillGradient(
+                matrix4f,
+                bufferBuilder,
+                tooltipX - 4,
+                tooltipY - 3,
+                tooltipX - 3,
+                tooltipY + tooltipHeight + 3,
+                zLevel,
+                backgroundColor,
+                backgroundColor);
+        fillGradient(
+                matrix4f,
+                bufferBuilder,
+                tooltipX + tooltipWidth + 3,
+                tooltipY - 3,
+                tooltipX + tooltipWidth + 4,
+                tooltipY + tooltipHeight + 3,
+                zLevel,
+                backgroundColor,
+                backgroundColor);
+        fillGradient(
+                matrix4f,
+                bufferBuilder,
+                tooltipX - 3,
+                tooltipY - 3 + 1,
+                tooltipX - 3 + 1,
+                tooltipY + tooltipHeight + 3 - 1,
+                zLevel,
+                borderColorStart,
+                borderColorEnd);
+        fillGradient(
+                matrix4f,
+                bufferBuilder,
+                tooltipX + tooltipWidth + 2,
+                tooltipY - 3 + 1,
+                tooltipX + tooltipWidth + 3,
+                tooltipY + tooltipHeight + 3 - 1,
+                zLevel,
+                borderColorStart,
+                borderColorEnd);
+        fillGradient(
+                matrix4f,
+                bufferBuilder,
+                tooltipX - 3,
+                tooltipY - 3,
+                tooltipX + tooltipWidth + 3,
+                tooltipY - 3 + 1,
+                zLevel,
+                borderColorStart,
+                borderColorStart);
+        fillGradient(
+                matrix4f,
+                bufferBuilder,
+                tooltipX - 3,
+                tooltipY + tooltipHeight + 2,
+                tooltipX + tooltipWidth + 3,
+                tooltipY + tooltipHeight + 3,
+                zLevel,
+                borderColorEnd,
+                borderColorEnd);
+        RenderSystem.enableDepthTest();
+        RenderSystem.disableTexture();
+        RenderSystem.enableBlend();
+        RenderSystem.defaultBlendFunc();
+        bufferBuilder.end();
+        BufferUploader.end(bufferBuilder);
+        RenderSystem.disableBlend();
+        RenderSystem.enableTexture();
+
+        // text
+        MultiBufferSource.BufferSource bufferSource =
+                MultiBufferSource.immediate(Tesselator.getInstance().getBuilder());
+        poseStack.translate(0.0, 0.0, 400.0);
+        int s = tooltipY;
+        boolean first = true;
+        for (ClientTooltipComponent line : lines) {
+            line.renderText(font, tooltipX, s, matrix4f, bufferSource);
+            s += line.getHeight() + (first ? 2 : 0);
+            first = false;
+        }
+        bufferSource.endBatch();
+        poseStack.popPose();
+    }
+
+    public static void copyImageToClipboard(BufferedImage bi) {
+        class ClipboardImage implements Transferable {
+            Image image;
+
+            public ClipboardImage(Image image) {
+                this.image = image;
+            }
+
+            @Override
+            public DataFlavor[] getTransferDataFlavors() {
+                return new DataFlavor[] {DataFlavor.imageFlavor};
+            }
+
+            @Override
+            public boolean isDataFlavorSupported(DataFlavor flavor) {
+                return DataFlavor.imageFlavor.equals(flavor);
+            }
+
+            @Override
+            public Object getTransferData(DataFlavor flavor) throws UnsupportedFlavorException {
+                if (!DataFlavor.imageFlavor.equals(flavor)) throw new UnsupportedFlavorException(flavor);
+                return this.image;
+            }
+        }
+
+        ClipboardImage ci = new ClipboardImage(bi);
+        Toolkit.getDefaultToolkit().getSystemClipboard().setContents(ci, null);
+    }
+
+    public static BufferedImage createScreenshot(RenderTarget fb) {
+        NativeImage image = new NativeImage(fb.width, fb.height, false);
+        RenderSystem.bindTexture(fb.getColorTextureId());
+        image.downloadTexture(0, false);
+        image.flipY();
+
+        int[] pixelValues = image.makePixelArray();
+        BufferedImage bufferedimage = new BufferedImage(fb.width, fb.height, BufferedImage.TYPE_INT_ARGB);
+        bufferedimage.setRGB(0, 0, fb.width, fb.height, pixelValues, 0, fb.width);
+        return bufferedimage;
     }
 }
