+++ common/src/main/java/com/wynntils/wc/utils/EntityUtils.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright © Wynntils 2022.
+ * This file is released under AGPLv3. See LICENSE for full license details.
+ */
+package com.wynntils.wc.utils;
+
+import com.wynntils.mc.utils.McUtils;
+import java.util.List;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.entity.animal.horse.AbstractHorse;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.phys.AABB;
+
+public class EntityUtils {
+
+    public static boolean isPlayersHorse(AbstractHorse horse, Player player) {
+        if (horse == null) return false;
+
+        String playerName = player.getName().getString();
+        String defaultName = "§f" + playerName + "§7" + "'s horse";
+        String customNameSuffix = "§7" + " [" + playerName + "]";
+
+        Component horseName = horse.getCustomName();
+        if (horseName == null) return false;
+
+        return defaultName.equals(horseName.getString())
+                || horseName.getString().endsWith(customNameSuffix);
+    }
+
+    public static AbstractHorse searchForHorseNearby(int searchRadius, Player player) {
+        List<AbstractHorse> horses = McUtils.mc()
+                .level
+                .getEntitiesOfClass(
+                        AbstractHorse.class,
+                        new AABB(
+                                player.getX() - searchRadius,
+                                player.getY() - searchRadius,
+                                player.getZ() - searchRadius,
+                                player.getX() + searchRadius,
+                                player.getY() + searchRadius,
+                                player.getZ() + searchRadius));
+
+        return horses.stream()
+                .filter(horse -> isPlayersHorse(horse, player))
+                .findFirst()
+                .orElse(null);
+    }
+}
